add ingredients into Pasta cooked

+++ db/migrate/20230707132936_create_preps.rb
@@ -1,7 +1,7 @@
 class CreatePreps < ActiveRecord::Migration[7.0]
   def change
     create_table :preps do |t|
-      t.integer :portion
+      t.float :portion
       t.references :ingredient, null: false, foreign_key: true
       t.references :recipe, null: false, foreign_key: true
